Camera

Use calibrated cameras in OpenCV convention:

The camera looks along the positive Z-axis, with X → right and Y → down.

The resulting image has its origin in the top-left with: X → right and Y → down.

SciGL.Camera — Type

Camera(cv_camera, [orthographic_camera=OrthographicCamera(cv_camera)])

Creates a SceneObject{Camera} which contains the projection matrix of the cvcamera. Optionally a custom orthographiccamera can be provided, e.g. for cropping views.

SciGL.Camera — Type

Camera

Strongly typed camera type which contains a static projection matrix.

SciGL.CvCamera — Type

CvCamera

A Camera parametrized like OpenCV. The convention is as in OpenCV: x-right, y-down, z-forward. Construct a Camera from it to be used in the shaders

SciGL.CvCamera — Method

CvCamera(width, height, f_x, f_y, c_x, c_y; [s=0, distortion=zeros(8), near=0.01, far=100])

Constructor with reasonable defaults

Parameters

width horizontal resolution [pixel]

height: vertical resolution in [pixel]

f_x: horizontal focal length [pixel/m]

f_y: vertical focal length [pixel/m]

c_x: horizontal center offset [pixel]

c_y: vertical center offset [pixel]

s: axis skew

distortion: distortion coefficients

near: near plane for OpenGL

far: far plane for OpenGL

SciGL.OrthographicCamera — Type

GLOrthographicCamera

Parametrizes an OrthographicCamera which transforms a cuboid space with the given parameters ([left,bottom,near],[right,top,far]) to normalized device coordinates ([-1,-1,-1],[1,1,1]).

SciGL.OrthographicCamera — Method

OrthographicCamera(c::CvCamera)

Extracts the orthographic scaling from the OpenCV camera. Since the origin in OpenGL is in the bottom-left and in OpenCV in the top-left, images will appear upside down in the OpenGL window but upright in memory.

SciGL.crop — Method

crop(cv_camera, left, right, top, bottom)

Creates a SceneObject{Camera} which contains the projection matrix of the cvcamera. This camera does not render the full size image of the cvcamera but only the area described by the bounding box (1, 1, width, height) → (left, right, top, bottom). Uses Julia indices starting at 1 and including the right & bottom, like img[left:right,top:bottom]. Does not clamp the values since an image can still be rendered.

SciGL.crop_size — Method

crop_size(left, right, top, bottom)

Size of the corresponding to img[left:right,top:bottom]

SciGL.lookat — Function

lookat(camera_t, object_t, [up=[0,-1.0]])

Calculates the Rotation to look at the object along positive Z with up defining the upwards direction. Default is the OpenCV convention: up = negative Y.

SciGL.orthographic_matrix — Method

orthographic_matrix(c::GLOrthoCamera)

Calculates the orthographic projection matrix like glOrtho

SciGL.perspective_matrix — Method

perspective_matrix(cv_camera)

Generate the perspective transformation matrix from the OpenCV camera parameters. Takes care of the OpenGL vs. OpenCV convention:

looking down negative Z vs. positive Z

up in image is positive Y vs. negative Y

Thus, use the OpenCV convention in following steps.

SciGL.to_gpu — Method

to_gpu(program, scene_camera)

Transfers the view and projection matrices to the OpenGL program

SciGL.view_matrix — Method

view_matrix(so::SceneObject{<:AbstractCamera})

Calculates the view matrix for a camera pose. The convention is as in OpenCV: x-rigth, y-down, z-forward