OffscreenContext

Simple entrypoint for rendering and transferring to CPU / CUDA.

SciGL.OffscreenContext — Type

OffscreenContext

Keep everything required for offscreen rendering and transfer to a (Cu)Array in one place.

High level API draw(context, scenes...): Synchronously draws and transfers the scenes to the (Cu)Array. During construction the context's framebuffer is bound once. Make sure to bind it again if you unbind it.

Low level API might be useful if you have calculations to execute during the transfer:

draw_framebuffer draws a scene to the texture attachment of the framebuffer

start_transfer to asynchronously transfer framebuffer → gl_buffer / render_data

wait_transfer to wait for the transfer to finish

Create a view via @view render_data[:, :, 1:n_images] which now contains the n_images renderings.

WARN Does not obey Julia image conventions but returns (x,y,z) instead (y,x,z) coordinates. PermutedDimsArrays are not CUDA compatible due to scalar indexing and transposing via permutedims is quite expensive.

SciGL.color_offscreen_context — Method

color_offscreen_context(width, height, [depth=1, array_type=Array, shaders=(SimpleVert, NormalFrag)])

Simplified generation of an OpenGL context for rendering images of a specific size. Batched rendering is enabled by generating a 3D Texture of RGBA pixels with size (width, height, depth). Specify the array_type as Array or CuArray.

SciGL.depth_offscreen_context — Method

depth_offscreen_context(width, height, [depth=1, array_type=Array])

Simplified generation of an OpenGL context for rendering depth images of a specific size. Batched rendering is enabled by generating a 3D Texture of Float32 with size (width, height, depth). Specify the array_type as Array or CuArray.

The resulting OffscreenContext's render_data has the array_type{Float32} which allows seamless a transfer of the depth values from the texture.

SciGL.distance_offscreen_context — Method

distance_offscreen_context(width, height, [depth=1, array_type=Array])

Simplified generation of an OpenGL context for rendering distance maps of a specific size. While depth images simply contain the z coordinate, distance maps contain the euclidean distance of the point from the camera. Batched rendering is enabled by generating a 3D Texture of Float32 with size (width, height, depth). Specify the array_type as Array or CuArray.

The resulting OffscreenContext's render_data has the array_type{Float32} which allows seamless a transfer of the distance values from the texture.

SciGL.draw — Method

draw(context, scenes)

Synchronously draw the scenes into the layers of the contxt's framebuffer and transfer it to the render_data of the context. Returns a view of of size (width, height, length(scenes)).

WARN: Overwrites the data in the context, copy it if you need it to persist!

SciGL.draw — Method

draw(context, scene)

Synchronously transfer the image with the given depth from OpenGL to the render_data. Returns a view of of size (width, height).

SciGL.draw_framebuffer — Function

draw_framebuffer(context, scene, [layer_id=1])

Render the scene to the framebuffer of the context.

SciGL.start_transfer — Function

start_transfer(context, [depth=1])

Start the asynchronous transfer the image with the given depth from OpenGL to the render_data.

WARN: Overwrites the data in the context, copy it if you need it to persist!

SciGL.to_device — Method

to_device(context, A)

Transfer A to the device of the render data, e.g. Array or CuArray

SciGL.transfer — Method

transfer(context, depth)

Synchronously transfer the image from OpenGL to the render_data. Returns a view of of size (width, height, depth).

WARN: Overwrites the data in the context, copy it if you need it to persist!

SciGL.transfer — Method

transfer(context)

Synchronously transfer the image from OpenGL to the render_data. Returns a view of of size (width, height).

WARN: Overwrites the data in the context, copy it if you need it to persist!

SciGL.wait_transfer — Function

start_transfer(context, [timeout_ns=10])

Wait for the transfer started in start_transfer to finish. Allows to set the timeout in nanoseconds of the glClientWaitSync call in the sync loop.

Context Creation

Manually create and destroy standalone contexts. Compared to the OffscreenContext batteries are not included, i.e. no textures to render to.

SciGL.clear_buffers — Method

clear_buffers()

Clears color, depth, and stencil.

SciGL.context_fullscreen — Method

context_fullscreen(width, height; name, window_hints)

Create an OpenGL context in fullscreen mode and makes it current.

SciGL.context_offscreen — Method

context_offscreen(width, height; name, window_hints)

Create an OpenGL context which is not visible, e.g. for offscreen rendering and makes it current.

SciGL.context_window — Method

context_window(width, height; name, window_hints)

Create an OpenGL context in windowed mode and makes it current.

SciGL.destroy_context — Method

destroy_context(context)

Generalization of GLFW.DestroyWindow

SciGL.enable_depth_stencil — Method

enable_depth_stencil()

Enable depth and stencil testing

SciGL.set_clear_color — Function

set_clear_color(color)

Set a color which is used for glClear(GLCOLORBUFFER_BIT), default is black

SciGL.set_context! — Method

set_context!(window)

Use this context for the current thread.

Layered Rendering

Low level interface to switch layers.

SciGL.activate_layer — Method

activate_layer(frambeuffer, id)

Activates layer of the first attachement to framebuffer. Does not for RBOs since they are inherently 2D.