Transformations

Common objects and transformation functions for 3D transformations.

CoordinateTransformations.AffineMap — Method

SMatrix(p::pose, s::Scale)

Creates the active transformation by first scaling, then rotating and finally translating

CoordinateTransformations.AffineMap — Method

AffineMap(p::Pose)

Creates the active transformation by first rotating and then translating

StaticArraysCore.SMatrix — Method

SMatrix(a::AffineMap)

Converts an AffineMap to a static affine transformation matrix.

StaticArraysCore.SMatrix — Method

SMatrix(p::pose, s::Scale)

Converts a Pose and Scale to an affine transformation matrix. Convention: (t ∘ r ∘ s)(x)

StaticArraysCore.SMatrix — Method

SMatrix(p::pose)

Converts a Pose to an affine transformation matrix. Convention: (t ∘ r)(x)

SciGL.augmented_matrix — Method

augmented_matrix(M, v)

Converts a linear map and a translation vector to an augmented affine transformation matrix. The matrix is of type Float32 and row major for OpenGL.

SciGL.decompose — Method

decompose(a::AbstractAffineMap)

Extract the linear map and translation as matrix and vector. OpenGL requires element types of Float32.